test(step2): test 분리
