+++ src/test/java/roomescape/theme/ThemeReadTest.java
@@ -1,0 +1,58 @@
+package roomescape.theme;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpStatus;
+import org.springframework.test.annotation.DirtiesContext;
+import roomescape.theme.application.ThemeService;
+import roomescape.theme.ui.dto.ThemeRequest;
+import roomescape.theme.ui.dto.ThemeResponse;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT, properties = {"server.port=8888"})
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class ThemeReadTest {
+    @Autowired
+    private ThemeService themeService;
+
+    @BeforeEach
+    public void setPort() {
+        RestAssured.port = 8888;
+    }
+
+    @Test
+    @DisplayName("전체 테마 조회")
+    void 전체_테마_조회() {
+        String name = "수키도키";
+        String description = "흐르는 대로 살자 해파리처럼🪼";
+        String thumbnail = "https://pbs.twimg.com/media/GApx6fjagAAkFsX.jpg";
+        themeService.add(ThemeRequest.create(name, description, thumbnail));
+
+        var response = RestAssured
+                .given().log().all()
+                .when().get("/themes")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract();
+
+        assertThat(response.jsonPath().getList("", ThemeResponse.class)).hasSize(1);
+    }
+
+    @Test
+    @DisplayName("테마가 하나도 없는 경우 전체 테마 조회")
+    void 등록된_테마_없는_경우_전체_테마_조회() {
+        var response = RestAssured
+                .given().log().all()
+                .when().get("/themes")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract();
+
+        assertThat(response.jsonPath().getList("", ThemeResponse.class)).hasSize(0);
+    }
+}
